+++ chevah/empirical/nose_test_timer.py
@@ -1,33 +1,6 @@
-"""This plugin provides test timings to identify which tests might be
-taking the most. From this information, it might be useful to couple
-individual tests nose's `--with-profile` option to profile problematic
-tests.
-
-This plugin is heavily influenced by nose's `xunit` plugin.
-
-Add this command to the way you execute nose::
-
-    --with-test-timer
-
-After all tests are executed, they will be sorted in ascending order.
-
-(c) 2011 - Mahmoud Abdelkader (http://github.com/mahmoudimus)
-
-LICENSE:
-            DO WHAT THE FUCK YOU WANT TO PUBLIC LICENSE
-                    Version 2, December 2004
-
- Copyright (C) 2004 Sam Hocevar <sam@hocevar.net>
-
- Everyone is permitted to copy and distribute verbatim or modified
- copies of this license document, and changing it is allowed as long
- as the name is changed.
-
-            DO WHAT THE FUCK YOU WANT TO PUBLIC LICENSE
-   TERMS AND CONDITIONS FOR COPYING, DISTRIBUTION AND MODIFICATION
-
-  0. You just DO WHAT THE FUCK YOU WANT TO.
-
+"""
+This plugin provides test timings to identify which tests might be
+taking the most.
 """
 
 import operator
@@ -132,7 +105,6 @@ class TestTimer(Plugin):
             total_time['outer'] += self._timed_tests['outer'][test]
 
         if tests_count == 0:
-            stream.writeln('-' * 70)
             stream.writeln('No tests were executed.')
             return
 
@@ -140,7 +112,10 @@ class TestTimer(Plugin):
         if test_start < 0:
             test_start = 0
 
-        for index in xrange(test_start, tests_count):
+        stream.writeln('-' * 70)
+        stream.writeln('Time usage top %s report:\n' % (TOP_COUNT))
+
+        for index in xrange(tests_count - 1, test_start - 1, -1):
             test, time_taken = inner_time_list[index]
             outer_time = self._timed_tests['outer'][test]
             stream.writeln("%0.4f:%0.4f: %s" % (time_taken, outer_time, test))